fix toolbar
